# api/User-openapi.yaml
openapi: 3.1.0
info:
  title: User Service Papich API
  description: "Документация API для микросервиса юзеров Папич. Управление юзерами\
    , ролями и сессиями"
  version: 1.0.0
servers:
  - url: http://localhost:8081
    description: Generated server url
...
paths:
...
  /sessions:
    post:
      tags:
        - Сессии
      summary: Войти в аккаунт
      description: Создает новую сессию юзера
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionLoginDto"
        required: true
      responses:
        "200":
          description: Успешный вход в аккаунт
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponseDto"
        "401":
          description: Указан неверный логин или пароль
        "500":
          description: Неизвестная ошибка сервера
        "400":
          description: "Неверный запрос от клиента, нарушение ограничений запроса\
            \ (тело, параметры)"
  /sessions/auth/token:
    get:
      tags:
        - Сессии
      summary: Обновить access токен
      description: Обновляет access и refresh токены для сессии по refresh токену
      operationId: refreshToken
      parameters:
        - name: refreshToken
          in: cookie
          description: Refresh токен
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Успешное обновление токенов
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionRefreshResponse"
        "401":
          description: Запрос не авторизован (refresh токен протух или неверный)
        "403":
          description: Сессия уже протухла
        "404":
          description: Сессия не найдена
        "500":
          description: Неизвестная ошибка сервера
        "400":
          description: "Неверный запрос от клиента, нарушение ограничений запроса\
            \ (тело, параметры)"
    post:
      tags:
        - Сессии
      summary: Обновить access токен
      description: Обновляет access и refresh токены для сессии по refresh токену
      operationId: refreshTokenPost
      parameters:
        - name: refreshToken
          in: cookie
          description: Refresh токен
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Успешное обновление токенов
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionRefreshResponse"
        "401":
          description: Запрос не авторизован (refresh токен протух или неверный)
        "403":
          description: Сессия уже протухла
        "404":
          description: Сессия не найдена
        "500":
          description: Неизвестная ошибка сервера
        "400":
          description: "Неверный запрос от клиента, нарушение ограничений запроса\
            \ (тело, параметры)"
...
  /sessions/logout/{sessionId}:
    patch:
      tags:
        - Сессии
      summary: Выйти из аккаунта
      description: "Помечает сессию неактивной, удаляет refresh токен"
      operationId: logout
      parameters:
        - name: sessionId
          in: path
          description: ID сессии
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Успешный выход из аккаунта
        "401":
          description: Запрос не авторизован
        "403":
          description: Попытка выхода из чужой сессии
        "404":
          description: Сессия не найдена
        "500":
          description: Неизвестная ошибка сервера
        "400":
          description: "Неверный запрос от клиента, нарушение ограничений запроса\
            \ (тело, параметры)"
      security:
        - bearerAuth: [ ]
...
components:
  schemas:
...
    SessionLoginDto:
      type: object
      description: Тело запроса для входа в аккаунт
      properties:
        email:
          type: string
          format: email
          description: Имейл
          minLength: 1
        password:
          type: string
          description: Пароль
          minLength: 1
      required:
        - email
        - password
    SessionResponseDto:
      type: object
      description: Ответ с инфой о сессии
      properties:
        id:
          type: string
          format: uuid
          description: ID сессии
        accessToken:
          type: string
          description: Access токен
    SessionRefreshResponse:
      type: object
      description: Ответ с обновленным access токеном
      properties:
        token:
          type: string
          description: Access токен
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT